update build script
For some reason the build script was listing the `LibraryProduct` three times.  It was also calling `autobuild` rather than `build_tarballs`, and manually doing a bunch of the things that `build_tarballs` does for you — was this from a really old version of the wizard?

It also might be nicer to use `platforms = supported_platforms()` and then `filter!` out any platforms that don't work (FreeBSD, I guess?).